Split Settings into several files
